Initial commit
Development history isn't relevant at this point as this is first
working version of oathtool.sh wrapper.

Script is both bash and dash compliant. Thus /bin/sh is used in shebang.

External utility dependencies are described in "Requires:" tag at the
beginning of the script.

Tested and found to be working with in both GPG and non-GPG setups on
Ubuntu 16.04 LTS and CentOS 7.x.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>